test: add helpers.go for shared test utilities
